requirements.txt
# Django and Core Dependencies
django>=5.2.9
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Database
psycopg2-binary>=2.9.9  # PostgreSQL adapter

# Cache and Session
django-redis>=5.4.0
redis>=5.0.1

# Environment Variables
python-decouple>=3.8

# Static Files
whitenoise>=6.6.0

# WSGI Server
gunicorn>=21.2.0

# Security
cryptography>=41.0.7

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
